added leave route and refactored
