+++ routes/club.js
@@ -1,8 +1,8 @@
 const Router = require('express');
 const {
-  getAll, deleteOne, getOne, createOne, join,
-} = require('../../controllers/club');
-const authenticate = require('../../middleware/auth');
+  getAll, deleteOne, getOne, createOne, join, leave,
+} = require('../controllers/club');
+const authenticate = require('../middleware/auth');
 
 const router = Router();
 
@@ -16,5 +16,8 @@ router.route('/:id')
 
 router.route('/join/:id')
   .put(authenticate, join());
+
+router.route('/leave/:id')
+  .put(authenticate, leave());
 
 module.exports = router;
